Switched to document listener extension point
It is better to use an extension point rather than listening
events from the event bus. It is safer (see https://youtrack.jetbrains.com/issue/IDEA-71701)
and improved readability.
